Assured that all values for the knobs are shown with the correct number format. And fixed the crash upon modifying the arpeggio durations.

+++ WaveSabreCore/include/WaveSabreCore/Pandora.h
@@ -274,8 +274,8 @@ namespace WaveSabreCore
 		// Arpeggio:
 		ArpeggioType arpeggioType;
 		int arpeggioNumOctaves;
-		int arpeggioInterval;
-		int arpeggioNoteDuration;
+		__int64 arpeggioInterval;
+		__int64 arpeggioNoteDuration;
 	};
 }
